permission cache

+++ app/Traits/ActivityTraits.js
@@ -18,30 +18,30 @@ class ActivityTraits {
 
   async saveActivity(request, auth, activity) {
     const NODE_ENV = Env.get("NODE_ENV")
-    // if (NODE_ENV === "production") {
-    try {
-      const headers = request.headers()
-      const user = await auth.getUser()
+    if (NODE_ENV === "production") {
+      try {
+        const headers = request.headers()
+        const user = await auth.getUser()
 
-      const activityData = {
-        user: {
-          id: user.id.toString(),
-          email: user.email,
-        },
-        browser: headers["user-agent"],
-        activity,
-        ip: request.ip(),
+        const activityData = {
+          user: {
+            id: user.id.toString(),
+            email: user.email,
+          },
+          browser: headers["user-agent"],
+          activity,
+          ip: request.ip(),
+        }
+
+        await axios.post("/activities", activityData)
+        RedisHelper.delete("Activity_*")
+
+        return true
+      } catch (e) {
+        console.log(JSON.stringify(e))
+        ErrorLog(request, e)
       }
-
-      await axios.post("/activities", activityData)
-      RedisHelper.delete("Activity_*")
-
-      return true
-    } catch (e) {
-      console.log(JSON.stringify(e))
-      ErrorLog(request, e)
     }
-    // }
   }
 }
